rag_module/requirements.txt
# Core dependencies
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-core>=0.1.0
langchain-community>=0.0.20

# Vector database
chromadb>=0.4.0

# Data validation
pydantic>=2.0.0

# Environment variable management
python-dotenv>=1.0.0

# Testing dependencies
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# Optional: For better async support
aiohttp>=3.9.0

rank-bm25>=0.2.2       
scikit-learn>=1.3.0        
numpy>=1.24.0             
